added translate to support ticket

+++ src/handlers/support.ts
@@ -3,6 +3,7 @@ import DataBase from "../db";
 import Components from "../embedsAndComps/components";
 import Embeds from "../embedsAndComps/Embeds";
 import { SupportTicket } from "../types";
+import LanguageHandler from "./language";
 import RankHandler, { Rank } from "./rank";
 import SetupHanlder from "./setup";
 
@@ -30,19 +31,31 @@ class SupportTicketHandler {
     }
 
     async createTicket() {
+        const lang = (await DataBase.guildsCollection.findOne({ guildId: this.interaction.guildId }))?.language || "en";
         if (this.supportTicket.memberId) {
-            await this.interaction.reply({ content: "You already have an open ticket, you have to close it first.", ephemeral: true });
+            await this.interaction.reply({ content: LanguageHandler.getMessageByLang("alreadyHasTicket", lang), ephemeral: true });
         } else {
-            const supportCatagory = await this.guild.channels.fetch((await SetupHanlder.getConfigObject(this.guild.id)).supportCatagory) as CategoryChannel;
-            const supportTicketChannel = await supportCatagory.createChannel(`support ${this.interaction.user.username}`, { type: "GUILD_TEXT" });
-            this.supportTicket = { channelId: supportTicketChannel.id, guildId: this.interaction.guildId as string, deleted: false, memberId: this.interaction.user.id };
-            const ticketNumber = await DataBase.supportCollection.countDocuments({}) + 1;
-            await supportTicketChannel.send({ embeds: [Embeds.supportTicketMainMessage(ticketNumber, this.interaction.member as GuildMember)], components: [await Components.supportTicket()] })
-            const rankHandler = await RankHandler.createHandler(this.interaction.member as GuildMember);
-            await (await supportTicketChannel.send(`${rankHandler.getRank(Rank.SUPERVISOR)}`)).delete();
-            await (await supportTicketChannel.send(`${this.interaction.user}`)).delete();
-            await supportTicketChannel.permissionOverwrites.create(this.interaction.member as GuildMember, { VIEW_CHANNEL: true });
-            await this.interaction.deferUpdate();
+            const supportCatagory = this.guild.channels.cache.get((await SetupHanlder.getConfigObject(this.guild.id)).supportCatagory) as CategoryChannel;
+            if (supportCatagory) {
+                const supportTicketChannel = await supportCatagory.createChannel(`support ${this.interaction.user.username}`, { type: "GUILD_TEXT" });
+                const rankHandler = await RankHandler.createHandler(this.interaction.member as GuildMember);
+                const ticketNumber = await DataBase.supportCollection.countDocuments({}) + 1;
+
+                this.supportTicket = { channelId: supportTicketChannel.id, guildId: this.interaction.guildId as string, deleted: false, memberId: this.interaction.user.id };
+
+                await supportTicketChannel.send({ embeds: [Embeds.supportTicketMainMessage(ticketNumber, this.interaction.member as GuildMember, lang)], components: [await Components.supportTicket(lang)] })
+                await (await supportTicketChannel.send(`${rankHandler.getRank(Rank.SUPERVISOR)}`)).delete();
+                await (await supportTicketChannel.send(`${this.interaction.user}`)).delete();
+
+                await supportTicketChannel.permissionOverwrites.create(this.interaction.member as GuildMember, { VIEW_CHANNEL: true });
+
+                await this.interaction.deferUpdate();
+                await this.save();
+            } else {
+                await this.interaction.reply({ content: LanguageHandler.getMessageByLang("supportCatagoryError", lang), ephemeral: true });
+                return;
+            }
+
         }
     }
